+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Added
+- Polygon.io for crypto
+
 ## [4.0.0] - 2021-06-29
 ### Added
 - `OnChain` feed collector